+++ RELEASE_NOTES.md
@@ -1,5 +1,10 @@
 # relation_engine folder
 
+## 2.0.0
+
+- The GTDB taxa loader now requires an archaea taxonomy file as well as a bacterial file and
+  the `input_file` key in the configuration has been renamed. 
+
 ## 1.0.0
 
 - First release with release notes and a version based on the current state of the codebase.